refactor: inline startup initialization into NahidaDesktop.init
