--- src/main/init.ts
@@ -1,49 +1,0 @@
-import { InitDB } from "@main/internal/db";
-import { supportsWindowsDesktopFeatures } from "@shared/platform";
-import { app, protocol } from "electron";
-import type { NahidaDesktop } from "./index";
-import { setting } from "./internal/db/schema";
-import { NahidaProtocolHandler } from "./internal/protocol";
-import { startServer } from "./server";
-
-export async function startInit(desktop: NahidaDesktop) {
-    if (desktop.initialized) return;
-
-    if (supportsWindowsDesktopFeatures(process.platform)) {
-        desktop.lib.native.startTracking();
-    }
-
-    // init db
-    await InitDB(desktop.lib.db);
-
-    // init lang
-    const lang = await desktop.lib.db.query.setting.findFirst({
-        where: (t, { eq }) => eq(t.key, "language"),
-    });
-    if (!lang) {
-        const locale = app.getLocale();
-        if (locale.startsWith("en"))
-            await desktop.lib.db.insert(setting).values({ key: "language", value: "en" });
-        else if (locale === "ko")
-            await desktop.lib.db.insert(setting).values({ key: "language", value: "ko" });
-        else if (locale.startsWith("zh"))
-            await desktop.lib.db.insert(setting).values({ key: "language", value: "zh" });
-        else await desktop.lib.db.insert(setting).values({ key: "language", value: "en" });
-    }
-
-    // make server
-    try {
-        await startServer();
-    } catch (error) {
-        desktop.logger.error(`Failed to start server on port 1027: ${error}`, "Server");
-        throw error;
-    }
-
-    // make tray
-    desktop.lib.tray.createTray();
-
-    // register custom protocol
-    protocol.handle("nahida", async (req) => await NahidaProtocolHandler(desktop, req));
-
-    desktop.initialized = true;
-}
